backend/requirements.txt
adal==1.2.7
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
azure-applicationinsights==0.1.1
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.12.0
azure-search-documents==11.3.0
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dataclasses-json==0.5.7
dnspython==2.4.0
exceptiongroup==1.1.2
Flask==2.2.3
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
importlib-metadata==6.5.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.146
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
msal==1.22.0
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
portalocker==2.7.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
tenacity==8.2.2
tqdm==4.65.0
typing-extensions==4.5.0
typing-inspect==0.8.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
